refactor: separated logic for post_bid_schema and update_bid_schema; updated tests
