+++ api/schemas/update_bid_schema.py
@@ -1,30 +1,26 @@
 from marshmallow import Schema, fields, post_load, validates, ValidationError
-from api.models.bid_model import BidModel
+from datetime import datetime
+from .links_schema import LinksSchema
 from .phase_schema import PhaseSchema
 from .feedback_schema import FeedbackSchema
+from ..models.status_enum import Status
 
 
-# Marshmallow schema for request body
-class BidRequestSchema(Schema):
-    tender = fields.Str(
-        required=True,
-        error_messages={"required": {"message": "Missing mandatory field"}},
-    )
-    client = fields.Str(
-        required=True,
-        error_messages={"required": {"message": "Missing mandatory field"}},
-    )
+# Marshmallow schema
+class UpdateBidSchema(Schema):
+    _id = fields.UUID(required=True)
+    tender = fields.Str(required=True)
+    client = fields.Str(required=True)
+    bid_date = fields.Date(required=True)
     alias = fields.Str()
-    bid_date = fields.Date(
-        format="%d-%m-%Y",
-        required=True,
-        error_messages={"required": {"message": "Missing mandatory field"}},
-    )
-    bid_folder_url = fields.URL()
-    was_successful = fields.Boolean()
+    bid_folder_url = fields.Str()
+    status = fields.Enum(Status, by_value=True, required=True)
+    links = fields.Nested(LinksSchema, required=True)
+    was_successful = fields.Bool(required=True)
     success = fields.List(fields.Nested(PhaseSchema))
     failed = fields.Nested(PhaseSchema)
     feedback = fields.Nested(FeedbackSchema)
+    last_updated = fields.DateTime(required=True)
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -62,7 +58,7 @@ class BidRequestSchema(Schema):
                 )
             success_phase_values.add(phase_value)
 
-    # Creates a Bid instance after processing
     @post_load
-    def makeBid(self, data, **kwargs):
-        return BidModel(**data)
+    def set_last_updated(self, data, **kwargs):
+        data["last_updated"] = datetime.now().isoformat()
+        return data
