+++ helpers/helpers.py
@@ -2,8 +2,8 @@ from flask import jsonify
 import uuid
 from datetime import datetime
 from werkzeug.exceptions import UnprocessableEntity
-from api.schemas.bid_schema import BidSchema
-from api.schemas.bid_request_schema import BidRequestSchema
+from api.schemas.update_bid_schema import UpdateBidSchema
+from api.schemas.post_bid_schema import PostBidSchema
 from api.schemas.bid_id_schema import BidIdSchema
 
 
@@ -41,9 +41,9 @@ def is_valid_isoformat(string):
 
 def validate_and_create_bid_document(request):
     # Process input and create data model
-    bid_document = BidRequestSchema().load(request)
+    bid_document = PostBidSchema().load(request)
     # Serialize to a JSON object
-    data = BidSchema().dump(bid_document)
+    data = PostBidSchema().dump(bid_document)
     return data
 
 
@@ -56,14 +56,19 @@ def validate_bid_id_path(bid_id):
 def validate_bid_update(user_request):
     if "status" in user_request:
         raise UnprocessableEntity("Cannot update status")
-    data = BidSchema().load(user_request, partial=True)
+    data = UpdateBidSchema().load(user_request, partial=True)
+    if "failed" in data:
+        data["failed"]["phase"] = data["failed"]["phase"].value
+    if "success" in data:
+        for obj in data["success"]:
+            obj["phase"] = obj["phase"].value
     return data
 
 
 def validate_status_update(user_request):
     if user_request == {}:
         raise UnprocessableEntity("Request must not be empty")
-    data = BidSchema().load(user_request, partial=True)
+    data = UpdateBidSchema().load(user_request, partial=True)
     if data:
         data["status"] = data["status"].value
     return data
