+++ request_examples/update_bid.http
@@ -1,6 +1,11 @@
-PUT http://localhost:8080/api/bids/7cea822f-fb27-4efd-87b3-3467eeb49d68 HTTP/1.1
+PUT http://localhost:8080/api/bids/0f05e112-6fd5-49b0-b7b0-e86aa376a61f HTTP/1.1
 Content-Type: application/json
 
 {
-  "status": "completed"
+  "failed": {
+    "has_score": true,
+    "out_of": 36,
+    "phase": 1,
+    "score": 28
+  }
 }
